Add directory API support
Most of the generic routines are ported from vitasdk, with platform specific bits modified to use Besta RTOS syscalls.

The directory name to find logic can be tweaked so it will be more fool proof (by e.g. checking for and filter out invalid FAT paths so it would be impossible to use a fnmatch pattern in diropen()) but these can be done later.

Closes #4.